backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose
sqlalchemy
psycopg2-binary
pydantic
bcrypt

